refactor(perf): split /research + /intelligence-map into server shell + motion island
Mirrors pattern from /music (W17), /golden-age (W18), /coaching (W19),
/ai-architecture (W20). Server handles metadata + JSON-LD + static layout.
Client island holds Framer Motion + interactive state.

- app/research: CollectionPage + ItemList + BreadcrumbList JSON-LD;
  client island keeps useState search/category filter and motion stagger.
- app/intelligence-map: TechArticle + BreadcrumbList JSON-LD;
  client island keeps useState node-expansion and AnimatePresence.

Server pages shrink to ~70 lines each; full motion + state lives in shells.
